requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
shap==0.44.0
cryptography==41.0.7
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
requests==2.32.5
Pillow==11.3.0
cryptography==41.0.7
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3
Pillow==11.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.32.5
scikit-learn==1.3.2
shap==0.44.0
whitenoise==6.5.0
